Frontend barzelim (#482)
* admin - invites refactor

* model changes

* apply flow - normal user

* admin - collector refactoring

* admin - fix approve collector

* cleanup mintClosed

* apply flow - trusted inviter

* mint - normal user

* mint honorary (wip) + fix "evolved" not updating

* cleanup mint view

* resolve minted and mintedHonorary

* CollectorLayout refactor

* invites page

* side menu - collector's room link and CTA

* format

* format smart-contracts

+++ smart-contracts/contracts/interface/IDiamondDawn.sol
@@ -6,5 +6,5 @@ interface IDiamondDawn {
 
     function mintHonorary(bytes calldata signature) external payable;
 
-    function safeEvolveCurrentPhase(uint tokenId) external payable;
+    function safeEvolveCurrentPhase(uint256 tokenId) external payable;
 }
