+++ smart-contracts/contracts/v1/interface/IDiamondDawnV1Mine.sol
@@ -5,30 +5,30 @@ import "../objects/Mine.sol";
 import "../objects/System.sol";
 
 interface IDiamondDawnV1Mine {
-    event Forge(uint tokenId);
-    event Mine(uint tokenId);
-    event Cut(uint tokenId);
-    event Polish(uint tokenId);
-    event Ship(uint tokenId, uint16 physicalId, uint64 number);
-    event Dawn(uint tokenId);
+    event Forge(uint256 tokenId);
+    event Mine(uint256 tokenId);
+    event Cut(uint256 tokenId);
+    event Polish(uint256 tokenId);
+    event Ship(uint256 tokenId, uint16 physicalId, uint64 number);
+    event Dawn(uint256 tokenId);
 
     function initialize(uint16 maxDiamond) external;
 
-    function forge(uint tokenId) external;
+    function forge(uint256 tokenId) external;
 
-    function mine(uint tokenId) external;
+    function mine(uint256 tokenId) external;
 
-    function cut(uint tokenId) external;
+    function cut(uint256 tokenId) external;
 
-    function polish(uint tokenId) external;
+    function polish(uint256 tokenId) external;
 
-    function ship(uint tokenId) external;
+    function ship(uint256 tokenId) external;
 
-    function dawn(uint tokenId) external;
+    function dawn(uint256 tokenId) external;
 
     function lockMine() external;
 
-    function getMetadata(uint tokenId) external view returns (string memory);
+    function getMetadata(uint256 tokenId) external view returns (string memory);
 
     function isReady(Stage stage) external view returns (bool);
 }
